Refactored security config, updated PGPropertyClient and application.yml
