+++ backend/sf-bookingmanagement/src/main/java/com/cdacproject/stayfinder/pg_booking_service/sf_bookingmanagement/client/PGPropertyClient.java
@@ -6,7 +6,7 @@ import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 
-@FeignClient(name = "SF-PGPROPERTY")
+@FeignClient(name = "SF-PROPERTY-SERVICE")
 public interface PGPropertyClient {
 
     @GetMapping("/api/pgs/{pgId}")
