+++ backend/sf-propertymanagement/src/main/resources/application.yml
@@ -2,6 +2,7 @@ spring:
   application:
     name: sf-property-service
 
+
   datasource:
     url: jdbc:postgresql://aws-0-ap-south-1.pooler.supabase.com:6543/postgres
     username: postgres.qwxlricixjwljrdmvecy
@@ -13,9 +14,6 @@ spring:
       ddl-auto: update
     show-sql: true
     database-platform: org.hibernate.dialect.PostgreSQLDialect
-
-jwt:
-  secret: your_very_long_secret_key_at_least_32_chars
 
 gateway:
   base-url: http://api-gateway:4000
